fix: Resolve compilation errors in PrivacyGuard, StorageManager, SystemCleaner
- PrivacyGuardFragment: Fix typo dangerPerPerms -> dangerPerms
- StorageManagerFragment: Add toInt() for Long to Int conversion
- SystemCleanerFragment: Add missing imports (ActivityManager, IntentFilter)
- SystemCleanerFragment: Add explicit type params for emptyList()
- SystemCleanerFragment: Fix Long/Int type mismatch in storage percentage

+++ app/src/main/java/com/hackerlauncher/modules/StorageManagerFragment.kt
@@ -173,7 +173,7 @@ class StorageManagerFragment : Fragment() {
 
             appendOutput("═══ FULL STORAGE ANALYSIS ═══\n\n")
             for ((name, size) in sorted) {
-                val pct = if (total > 0) (size * 100 / total) else 0
+                val pct = if (total > 0) ((size * 100 / total).toInt()) else 0
                 val bar = "█".repeat((pct / 3).coerceAtMost(25))
                 appendOutput("${name.take(15).padEnd(15)} ${formatSize(size).padEnd(10)} $bar $pct%\n")
             }
